refactor(userspace): extract route/neighbor snapshots to protocol_routes.go
Fifth code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the FIB route and manager-neighbor snapshots into a
cohesive sibling protocol_routes.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change): RouteSnapshot,
NeighborSnapshot.

Behavior byte-identical; json tags preserved for the Rust userspace-dp
deserializer (NeighborSnapshot rides the update_neighbors wire). Build
passes; gofmt clean.
